Prelease Changes: Check Description
Leaper: added the bonus movement into the tip image, slightly changed the logic for the flip + fire preview image.

Centipede: removed `icon_swap_acid_off_glowB`, just used the normal one and made a copy with a new offset

Psion: added custom tip images to all versions of the weapon, to better showcase the interactions it has, also added a counter to the `Diplomatic Immunity` achievement, so that you know how far you are to complete it. Added a mod option that allows you to pick the class of the Psionic Receiver, so it can be cyborg class, or have no class, like in vanilla.

All 3: their weapons will only be unlocked if their respective achievement is. changed internally how the sprites are added.

Shield Psion: changed the idle animation to one that's miles better, pretty proud of that one 😃